refactor: atom, molecules, organism 디렉토리 복수형으로 명칭 변경
