+++ src/components/atoms/main/header/AccountLogoStyles.ts
