+++ src/components/atoms/main/header/AccountTextStyles.ts
