+++ src/components/atoms/main/sidebar/MainLogo.tsx
