+++ src/components/atoms/main/sidebar/MainLogoStyles.ts
