+++ src/components/molecules/main/body/MainBodyContent.tsx
@@ -1,6 +1,6 @@
 import { MainBodyContentBox } from "./MainBodyContentStyles";
-import Title from "~/components/atom/main/body/Title";
-import Description from "~/components/atom/main/body/Description";
+import Title from "~/components/atoms/main/body/Title";
+import Description from "~/components/atoms/main/body/Description";
 
 const MainBodyContent = () => (
   <MainBodyContentBox>
