--- src/components/organisms/main/MainContentStyles.ts
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-import tw from "twin.macro";
-
-export const MainContentBox = styled.div`
-  ${tw`
-    flex
-    h-screen
-    overflow-x-hidden
-  `}
-`;
