+++ src/components/organisms/main/sidebar/SideBar.tsx
@@ -1,5 +1,5 @@
 import { SidebarBox } from "./SideBarStyles";
-import MainLogo from "~/components/atom/main/sidebar/MainLogo";
+import MainLogo from "~/components/atoms/main/sidebar/MainLogo";
 import SocialIcons from "~/components/molecules/main/sidebar/SocialIcons";
 
 const Sidebar = () => (
